+++ docs/Whats-New.md
@@ -423,6 +423,7 @@ New:
 - Promotion animation (by Fryone)
 - Allow different technos to share build limit in a group (by ststl & Ollerus)
 - Map events `604-605` for checking if a specific Techno enters in a cell (by FS-21)
+- Raise alert when technos are taking damage (by 航味麻酱)
 
 Vanilla fixes:
 - Allow AI to repair structures built from base nodes/trigger action 125/SW delivery in single player missions (by Trsdy)